+++ tests/resources/test_downloadable_files.py
@@ -41,8 +41,11 @@ trial_id_2 = "test-trial-2"
 upload_types = ["wes_bam", "cytof"]
 
 
-def setup_downloadable_files(cidc_api) -> Tuple[int, int]:
-    """Insert two downloadable files into the database."""
+def setup_downloadable_files(cidc_api) -> Tuple[int, int, int]:
+    """
+    Insert three downloadable files into the database.
+    One is clinical_data to test that it isn't included in cross-assay permissions.
+    """
     metadata_json = {
         "protocol_identifier": trial_id_1,
         "allowed_collection_event_names": [],
@@ -68,20 +71,24 @@ def setup_downloadable_files(cidc_api) -> Tuple[int, int]:
     cytof_file = make_file(
         trial_id_2, "cytof/.../analysis.zip", "cytof", "/cytof_analysis/analysis.zip"
     )
+    clinical_file = make_file(
+        trial_id_1, "clinical/.../file.csv", "clinical_data", "/clinical/."
+    )
 
     with cidc_api.app_context():
         trial_1.insert()
         trial_2.insert()
         wes_file.insert()
         cytof_file.insert()
+        clinical_file.insert()
 
-        return wes_file.id, cytof_file.id
+        return wes_file.id, cytof_file.id, clinical_file.id
 
 
 def test_list_downloadable_files(cidc_api, clean_db, monkeypatch):
     """Check that getting a list of files works as expected"""
     user_id = setup_user(cidc_api, monkeypatch)
-    file_id_1, file_id_2 = setup_downloadable_files(cidc_api)
+    file_id_1, file_id_2, clinical_file_id = setup_downloadable_files(cidc_api)
 
     client = cidc_api.test_client()
 
@@ -96,7 +103,7 @@ def test_list_downloadable_files(cidc_api, clean_db, monkeypatch):
     for kwargs in [
         {"trial_id": trial_id_1, "upload_type": upload_types[0]},
         {"upload_type": upload_types[0]},
-        {"trial_id": trial_id_1},
+        {"trial_id": trial_id_1},  # won't include clinical_data
     ]:
         # Give the user one permission
         with cidc_api.app_context():
@@ -123,9 +130,11 @@ def test_list_downloadable_files(cidc_api, clean_db, monkeypatch):
         make_role(user_id, role, cidc_api)
         res = client.get("/downloadable_files")
         assert res.status_code == 200
-        assert len(res.json["_items"]) == 2
-        assert res.json["_meta"]["total"] == 2
-        assert set([f["id"] for f in res.json["_items"]]) == set([file_id_1, file_id_2])
+        assert len(res.json["_items"]) == 3
+        assert res.json["_meta"]["total"] == 3
+        assert set([f["id"] for f in res.json["_items"]]) == set(
+            [file_id_1, file_id_2, clinical_file_id]
+        )
 
         # Admin filter queries include any files that fit the criteria
         res = client.get(
@@ -139,12 +148,13 @@ def test_list_downloadable_files(cidc_api, clean_db, monkeypatch):
     # Make sure it's possible to sort by file extension
     res = client.get(f"/downloadable_files?sort_field=file_ext&sort_direction=asc")
     assert res.status_code == 200
-    assert [f["file_ext"] for f in res.json["_items"]] == ["bam", "zip"]
+    assert [f["file_ext"] for f in res.json["_items"]] == ["bam", "csv", "zip"]
 
     # Make sure it's possible to sort by data category
     res = client.get(f"/downloadable_files?sort_field=data_category&sort_direction=asc")
     assert res.status_code == 200
     assert [f["data_category"] for f in res.json["_items"]] == [
+        "Clinical Data",
         "CyTOF|Analysis Results",
         "WES|Source",
     ]
@@ -153,7 +163,7 @@ def test_list_downloadable_files(cidc_api, clean_db, monkeypatch):
 def test_get_downloadable_file(cidc_api, clean_db, monkeypatch):
     """Check that getting a single file works as expected."""
     user_id = setup_user(cidc_api, monkeypatch)
-    file_id_1, file_id_2 = setup_downloadable_files(cidc_api)
+    file_id_1, file_id_2, clinical_file_id = setup_downloadable_files(cidc_api)
 
     client = cidc_api.test_client()
 
@@ -166,7 +176,7 @@ def test_get_downloadable_file(cidc_api, clean_db, monkeypatch):
         perm = Permissions(
             granted_to_user=user_id,
             trial_id=trial_id_1,
-            upload_type=upload_types[0],
+            upload_type=None,  # cross trial applies to file_id_1 but not clinical_file_id
             granted_by_user=user_id,
         )
         perm.insert()
@@ -175,6 +185,9 @@ def test_get_downloadable_file(cidc_api, clean_db, monkeypatch):
     res = client.get(f"/downloadable_files/{file_id_1}")
     assert res.status_code == 200
     assert res.json["id"] == file_id_1
+    # Non-admins canNOT get any other single file
+    res = client.get(f"/downloadable_files/{clinical_file_id}")
+    assert res.status_code == 401
 
     # Admins can get any file regardless of permissions
     make_admin(user_id, cidc_api)
@@ -190,7 +203,7 @@ def test_get_downloadable_file(cidc_api, clean_db, monkeypatch):
 def test_get_related_files(cidc_api, clean_db, monkeypatch):
     """Check that the related_files endpoint calls `get_related_files`"""
     user_id = setup_user(cidc_api, monkeypatch)
-    file_id_1, file_id_2 = setup_downloadable_files(cidc_api)
+    file_id_1, file_id_2, _ = setup_downloadable_files(cidc_api)
 
     client = cidc_api.test_client()
 
@@ -236,7 +249,7 @@ def test_get_related_files(cidc_api, clean_db, monkeypatch):
 def test_get_filelist(cidc_api, clean_db, monkeypatch):
     """Check that getting a filelist.tsv works as expected"""
     user_id = setup_user(cidc_api, monkeypatch)
-    file_id_1, file_id_2 = setup_downloadable_files(cidc_api)
+    file_id_1, file_id_2, clinical_file_id = setup_downloadable_files(cidc_api)
 
     client = cidc_api.test_client()
 
@@ -247,7 +260,7 @@ def test_get_filelist(cidc_api, clean_db, monkeypatch):
     assert res.status_code == 422
 
     # User has no permissions, so no files should be found
-    short_file_list = {"file_ids": [file_id_1, file_id_2]}
+    short_file_list = {"file_ids": [file_id_1, file_id_2, clinical_file_id]}
     res = client.post(url, json=short_file_list)
     assert res.status_code == 404
 
@@ -270,21 +283,37 @@ def test_get_filelist(cidc_api, clean_db, monkeypatch):
         f"gs://{GOOGLE_ACL_DATA_BUCKET}/{trial_id_1}/wes/.../reads_123.bam\t{trial_id_1}_wes_..._reads_123.bam\n"
     )
 
+    # Give the user a cross-assay permission
+    with cidc_api.app_context():
+        perm = Permissions(
+            granted_to_user=user_id,
+            trial_id=trial_id_1,
+            upload_type=None,
+            granted_by_user=user_id,
+        )
+        perm.insert()
+
+    # User has cross-assay, so the filelist should NOT contain the clinical file
+    res = client.post(url, json=short_file_list)
+    assert res.status_code == 200
+    assert "text/tsv" in res.headers["Content-Type"]
+    assert "filename=filelist.tsv" in res.headers["Content-Disposition"]
+    assert res.data.decode("utf-8") == (
+        f"gs://{GOOGLE_ACL_DATA_BUCKET}/{trial_id_1}/wes/.../reads_123.bam\t{trial_id_1}_wes_..._reads_123.bam\n"
+    )
+
     # Admins don't need permissions to get files
     make_admin(user_id, cidc_api)
     res = client.post(url, json=short_file_list)
     assert res.status_code == 200
 
-    assert res.data.decode("utf-8") in [
-        (
-            f"gs://{GOOGLE_ACL_DATA_BUCKET}/{trial_id_1}/wes/.../reads_123.bam\t{trial_id_1}_wes_..._reads_123.bam\n"
-            f"gs://{GOOGLE_ACL_DATA_BUCKET}/{trial_id_2}/cytof/.../analysis.zip\t{trial_id_2}_cytof_..._analysis.zip\n"
-        ),
-        (  # reversed
-            f"gs://{GOOGLE_ACL_DATA_BUCKET}/{trial_id_2}/cytof/.../analysis.zip\t{trial_id_2}_cytof_..._analysis.zip\n"
-            f"gs://{GOOGLE_ACL_DATA_BUCKET}/{trial_id_1}/wes/.../reads_123.bam\t{trial_id_1}_wes_..._reads_123.bam\n"
-        ),
-    ]
+    assert sorted(res.data.decode("utf-8").strip().split("\n")) == sorted(
+        [
+            f"gs://{GOOGLE_ACL_DATA_BUCKET}/{trial_id_1}/wes/.../reads_123.bam\t{trial_id_1}_wes_..._reads_123.bam",
+            f"gs://{GOOGLE_ACL_DATA_BUCKET}/{trial_id_2}/cytof/.../analysis.zip\t{trial_id_2}_cytof_..._analysis.zip",
+            f"gs://{GOOGLE_ACL_DATA_BUCKET}/{trial_id_1}/clinical/.../file.csv\t{trial_id_1}_clinical_..._file.csv",
+        ]
+    )
 
     # Clear inserted file records
     with cidc_api.app_context():
@@ -313,10 +342,11 @@ def test_get_filelist(cidc_api, clean_db, monkeypatch):
 
 def test_create_compressed_batch(cidc_api, clean_db, monkeypatch):
     user_id = setup_user(cidc_api, monkeypatch)
-    file_id_1, file_id_2 = setup_downloadable_files(cidc_api)
+    file_id_1, file_id_2, clinical_file_id = setup_downloadable_files(cidc_api)
     with cidc_api.app_context():
         url_1 = DownloadableFiles.find_by_id(file_id_1).object_url
         url_2 = DownloadableFiles.find_by_id(file_id_2).object_url
+        clinical_url = DownloadableFiles.find_by_id(clinical_file_id).object_url
 
     client = cidc_api.test_client()
 
@@ -327,7 +357,7 @@ def test_create_compressed_batch(cidc_api, clean_db, monkeypatch):
     assert res.status_code == 422
 
     # User has no permissions, so no files should be found
-    short_file_list = {"file_ids": [file_id_1, file_id_2]}
+    short_file_list = {"file_ids": [file_id_1, file_id_2, clinical_file_id]}
     res = client.post(url, json=short_file_list)
     assert res.status_code == 404
 
@@ -336,7 +366,7 @@ def test_create_compressed_batch(cidc_api, clean_db, monkeypatch):
         perm = Permissions(
             granted_to_user=user_id,
             trial_id=trial_id_1,
-            upload_type=upload_types[0],
+            upload_type=None,  # cross-assay should not include clinical_data
             granted_by_user=user_id,
         )
         perm.insert()
@@ -379,31 +409,34 @@ def test_create_compressed_batch(cidc_api, clean_db, monkeypatch):
     make_admin(user_id, cidc_api)
 
     # Admin has access to both files, but together they are too large
-    # 2 files of 51MB each is 102MB > 100MB
+    # 3 files of 51MB each is 153MB > 100MB
     res = client.post(url, json=short_file_list)
     assert res.status_code == 400
     assert "batch too large" in res.json["_error"]["message"]
     bucket.get_blob.assert_not_called()
     blob.upload_from_filename.assert_not_called()
 
-    # Decrease the size of one of the files and try again
-    # 51MB + 1B < 100MB
+    # Decrease the size of two of the files and try again
+    # 51MB + 2 * 1B < 100MB
     with cidc_api.app_context():
-        df = DownloadableFiles.find_by_id(file_id_1)
-        df.file_size_bytes = 1
-        df.update()
+        for df_id in (file_id_1, file_id_2):
+            df = DownloadableFiles.find_by_id(df_id)
+            df.file_size_bytes = 1
+            df.update()
 
     res = client.post(url, json=short_file_list)
     assert res.status_code == 200
     assert res.json == signed_url
     assert call(url_1) in bucket.get_blob.call_args_list
     assert call(url_2) in bucket.get_blob.call_args_list
+    assert call(clinical_url) in bucket.get_blob.call_args_list
     blob.upload_from_filename.assert_called_once()
 
     mock_logger.info.assert_called_once()
     assert "test@email.com" in mock_logger.info.call_args[0][0]
     assert url_1 in mock_logger.info.call_args[0][0]
     assert url_2 in mock_logger.info.call_args[0][0]
+    assert clinical_url in mock_logger.info.call_args[0][0]
 
 
 def test_get_filter_facets(cidc_api, clean_db, monkeypatch):
@@ -413,7 +446,7 @@ def test_get_filter_facets(cidc_api, clean_db, monkeypatch):
 
     client = cidc_api.test_client()
 
-    def check_facet_counts(facets, wes_count=0, cytof_count=0):
+    def check_facet_counts(facets, wes_count=0, cytof_count=0, clinical_count=0):
         for facet, subfacets in facets["Assay Type"].items():
             if not isinstance(subfacets, list):
                 subfacets = [subfacets]
@@ -424,49 +457,53 @@ def test_get_filter_facets(cidc_api, clean_db, monkeypatch):
                     assert subfacet["count"] == cytof_count
                 else:
                     assert subfacet["count"] == 0
+        for facet in facets["Clinical Type"]:
+            if facet == "Clinical Data":
+                assert facet["count"] == clinical_count
 
     # Non-admins' facets take into account their permissions
     with cidc_api.app_context():
         perm = Permissions(
             granted_to_user=user_id,
             trial_id=trial_id_1,
-            upload_type=upload_types[0],
+            upload_type=None,  # cross-assay should not include clinical_data
             granted_by_user=user_id,
         )
         perm.insert()
     res = client.get("/downloadable_files/filter_facets")
     assert res.status_code == 200
-    check_facet_counts(res.json["facets"], wes_count=1, cytof_count=0)
+    check_facet_counts(res.json["facets"], wes_count=1, cytof_count=0, clinical_count=0)
 
     # Admins' facets include all files, regardless of permissions
     make_admin(user_id, cidc_api)
     res = client.get("/downloadable_files/filter_facets")
     assert res.status_code == 200
     assert sorted(res.json["trial_ids"], key=lambda x: x["label"]) == sorted(
-        [{"label": trial_id_1, "count": 1}, {"label": trial_id_2, "count": 1}],
+        [{"label": trial_id_1, "count": 2}, {"label": trial_id_2, "count": 1}],
         key=lambda x: x["label"],
     )
-    check_facet_counts(res.json["facets"], wes_count=1, cytof_count=1)
+    check_facet_counts(res.json["facets"], wes_count=1, cytof_count=1, clinical_count=1)
 
     # Trial facets are governed by data category facets
     res = client.get("/downloadable_files/filter_facets?facets=Assay Type|WES|Germline")
     assert res.status_code == 200
     assert res.json["trial_ids"] == []
-    check_facet_counts(res.json["facets"], wes_count=1, cytof_count=1)
+    check_facet_counts(res.json["facets"], wes_count=1, cytof_count=1, clinical_count=1)
 
     # Data category facets are governed by trial facets
     res = client.get("/downloadable_files/filter_facets?trial_ids=someothertrial")
     assert res.status_code == 200
     res.json["trial_ids"] == []
-    check_facet_counts(res.json["facets"], wes_count=0, cytof_count=0)
+    check_facet_counts(res.json["facets"], wes_count=0, cytof_count=0, clinical_count=0)
 
 
 def test_get_download_url(cidc_api, clean_db, monkeypatch):
     """Check that generating a GCS signed URL works as expected"""
     user_id = setup_user(cidc_api, monkeypatch)
-    file_id, _ = setup_downloadable_files(cidc_api)
+    file_id, _, clinical_file_id = setup_downloadable_files(cidc_api)
     with cidc_api.app_context():
         file_url = DownloadableFiles.find_by_id(file_id).object_url
+        clinical_url = DownloadableFiles.find_by_id(clinical_file_id).object_url
 
     # mock logs
     mock_logger = MagicMock()
@@ -493,7 +530,7 @@ def test_get_download_url(cidc_api, clean_db, monkeypatch):
         perm = Permissions(
             granted_to_user=user_id,
             trial_id=trial_id_1,
-            upload_type=upload_types[0],
+            upload_type=None,  # cross-assay should not include clinical_data
             granted_by_user=user_id,
         )
         perm.insert()
@@ -509,6 +546,9 @@ def test_get_download_url(cidc_api, clean_db, monkeypatch):
     mock_logger.info.assert_called_once()
     assert "test@email.com" in mock_logger.info.call_args[0][0]
     assert file_url in mock_logger.info.call_args[0][0]
+    # cross-assay should not include clinical_data
+    res = client.get(f"/downloadable_files/download_url?id={clinical_file_id}")
+    assert res.status_code == 404
 
     # network viewers aren't allowed to get download urls
     make_role(user_id, CIDCRole.NETWORK_VIEWER.value, cidc_api)